Replace old scheduler/worker with new ones
